+++ packages/react-auth-strapi/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.54](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.53...@ttoss/react-auth-strapi@0.2.54) (2025-12-23)
+
+**Note:** Version bump only for package @ttoss/react-auth-strapi
+
+## [0.2.53](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.52...@ttoss/react-auth-strapi@0.2.53) (2025-12-22)
+
+**Note:** Version bump only for package @ttoss/react-auth-strapi
+
+## [0.2.52](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.51...@ttoss/react-auth-strapi@0.2.52) (2025-12-22)
+
+**Note:** Version bump only for package @ttoss/react-auth-strapi
+
 ## [0.2.51](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.50...@ttoss/react-auth-strapi@0.2.51) (2025-12-18)
 
 **Note:** Version bump only for package @ttoss/react-auth-strapi