+++ terezinha-farm/api/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.30](https://github.com/ttoss/ttoss/compare/@terezinha-farm/api@0.19.29...@terezinha-farm/api@0.19.30) (2025-12-28)
+
+**Note:** Version bump only for package @terezinha-farm/api
+
+## 0.19.29 (2025-12-27)
+
+**Note:** Version bump only for package @terezinha-farm/api
+
+## 0.19.28 (2025-12-27)
+
+**Note:** Version bump only for package @terezinha-farm/api
+
 ## 0.19.27 (2025-12-09)
 
 **Note:** Version bump only for package @terezinha-farm/api